src/api/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
drf-yasg==1.21.7
neo4j-driver==4.4.11
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
torch==2.0.1
transformers==4.33.2
python-dotenv==1.0.0 